+++ src/Traits/ImmutableComponent.php
@@ -86,8 +86,11 @@ trait ImmutableComponent
         };
 
         $res = preg_replace_callback($regexp, $encoder, $str);
+        if (null !== $res) {
+            return $res;
+        }
 
-        return $res;
+        return rawurlencode($str);
     }
 
     /**